move build_zip and build_tar to private
