+++ pkg/tests/archive_test.py
@@ -18,7 +18,7 @@ import tarfile
 import unittest
 
 from bazel_tools.tools.python.runfiles import runfiles
-from rules_pkg import archive
+from rules_pkg.private import archive
 from tests import compressor
