+++ pkg/tests/pkg_deb_test.py
@@ -23,7 +23,7 @@ import tarfile
 import unittest
 
 from bazel_tools.tools.python.runfiles import runfiles
-from rules_pkg import archive
+from rules_pkg.private import archive
 
 
 class DebInspect(object):
